+++ src/config/imap.php
@@ -108,6 +108,9 @@ return [
     |       Default TRUE
     |   -Soft fail
     |       Default FALSE - Set to TRUE if you want to ignore certain exception while fetching bulk messages
+    |   -RFC822
+    |       Default TRUE - Set to FALSE to prevent the usage of \imap_rfc822_parse_headers().
+    |                      See https://github.com/Webklex/php-imap/issues/115 for more information.
     |   -Message key identifier option
     |       You can choose between the following:
     |       'id'     - Use the MessageID as array key (default, might cause hickups with yahoo mail)
@@ -139,6 +142,7 @@ return [
         'fetch_body' => true,
         'fetch_flags' => true,
         'soft_fail' => false,
+        'rfc822' => true,
         'message_key' => 'list',
         'fetch_order' => 'asc',
         'dispositions' => ['attachment', 'inline'],